+++ src/main/java/com/example/fschool/model/po/Students.java
@@ -77,7 +77,7 @@ public class Students implements Serializable {
      * 0表示女，1表示男
      */
     @TableField("student_sex")
-    private Byte studentSex;
+    private Integer studentSex;
 
     /**
      * 成绩
